Enabling access to private windows #45
This access is required for tracking when a page has lost user attention because the user has switched to a private window. Note that we are not planning to collect data about what a user does in private windows for the news study. Our tentative thinking is that any collection of data about what a user does in private windows—in future studies—should require express study-specific consent. TBD whether even that will be permitted.